requirements.txt
sqlalchemy
asyncpg
aiogram==2.7
aiohttp==3.6.2
aioredis==1.3.1
tortoise-orm==0.16.12